Add bump-my-version configuration
This commit introduces the bump-my-version tool to the project.
It includes a `.bumpversion.toml` file with configurations for
managing the project's version, including settings for parsing,
serializing, and updating the version in `setup.py` and
`pyradios/__about__.py`. The configuration also includes settings
for creating tags and commits during the version bumping process.